Replace entityName hook with a function that reads from store.
