+++ app/client/src/pages/Editor/Explorer/JSActions/JSActionContextMenu.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useMemo, useState } from "react";
+import React, { useCallback, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import TreeDropdown from "pages/Editor/Explorer/TreeDropdown";
 import ContextMenuTrigger from "../ContextMenuTrigger";
@@ -8,8 +8,8 @@ import {
   deleteJSCollection,
 } from "actions/jsActionActions";
 import { ContextMenuPopoverModifiers } from "../helpers";
-import { groupBy, noop } from "lodash";
-import { useNewJSCollectionName } from "./helpers";
+import noop from "lodash/noop";
+import { getJSEntityName } from "./helpers";
 import { initExplorerEntityNameEdit } from "actions/explorerActions";
 import { ReduxActionTypes } from "@appsmith/constants/ReduxActionConstants";
 import { ENTITY_TYPE } from "entities/DataTree/dataTreeFactory";
@@ -24,10 +24,6 @@ import {
   createMessage,
 } from "@appsmith/constants/messages";
 import { getPageListAsOptions } from "selectors/entitiesSelector";
-import store from "store";
-import { selectJSCollections } from "selectors/editorSelectors";
-import { JSCollectionData } from "reducers/entityReducers/jsActionsReducer";
-import { getNextEntityName } from "utils/AppsmithUtils";
 
 type EntityContextMenuProps = {
   id: string;
@@ -36,7 +32,6 @@ type EntityContextMenuProps = {
   pageId: string;
 };
 export function JSCollectionEntityContextMenu(props: EntityContextMenuProps) {
-  const nextEntityName = useNewJSCollectionName();
   const [confirmDelete, setConfirmDelete] = useState(false);
   const dispatch = useDispatch();
 
@@ -54,32 +49,9 @@ export function JSCollectionEntityContextMenu(props: EntityContextMenuProps) {
     [],
   );
 
-  const getEntityName = useCallback(() => {
-    const state = store.getState();
-    const jsCollections = selectJSCollections(state);
-    return (
-      name: string,
-      destinationPageId: string,
-      isCopyOperation?: boolean,
-    ) => {
-      const groupedActions = groupBy(jsCollections, "config.pageId");
-      const pageActions = groupedActions[destinationPageId] || [];
-      const actionNames = pageActions.map(
-        (action: JSCollectionData) => action.config.name,
-      );
-      return actionNames.indexOf(name) > -1
-        ? getNextEntityName(
-            isCopyOperation ? `${name}Copy` : name,
-            actionNames,
-            true,
-          )
-        : name;
-    };
-  }, []);
-
   const copyJSCollectionToPage = useCallback(
     (actionId: string, actionName: string, pageId: string) => {
-      const nextEntityName = getEntityName();
+      const nextEntityName = getJSEntityName();
       dispatch(
         copyJSCollectionRequest({
           id: actionId,
@@ -88,11 +60,11 @@ export function JSCollectionEntityContextMenu(props: EntityContextMenuProps) {
         }),
       );
     },
-    [dispatch, nextEntityName],
+    [dispatch],
   );
   const moveJSCollectionToPage = useCallback(
     (actionId: string, actionName: string, destinationPageId: string) => {
-      const nextEntityName = getEntityName();
+      const nextEntityName = getJSEntityName();
       dispatch(
         moveJSCollectionRequest({
           id: actionId,
@@ -101,7 +73,7 @@ export function JSCollectionEntityContextMenu(props: EntityContextMenuProps) {
         }),
       );
     },
-    [dispatch, nextEntityName, props.pageId],
+    [dispatch, props.pageId],
   );
   const deleteJSCollectionFromPage = useCallback(
     (actionId: string, actionName: string) =>
