+++ app/client/src/pages/Editor/Explorer/JSActions/MoreJSActionsMenu.tsx
@@ -1,18 +1,15 @@
 import React, { useCallback, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-
 import { AppState } from "reducers";
-
 import {
   moveJSCollectionRequest,
   copyJSCollectionRequest,
   deleteJSCollection,
 } from "actions/jsActionActions";
-
 import { ContextMenuPopoverModifiers } from "../helpers";
 import { noop } from "lodash";
 import TreeDropdown from "pages/Editor/Explorer/TreeDropdown";
-import { useNewJSCollectionName } from "./helpers";
+import { getJSEntityName } from "./helpers";
 import styled from "styled-components";
 import Icon, { IconSize } from "components/ads/Icon";
 import { Position } from "@blueprintjs/core";
@@ -70,31 +67,35 @@ export const MoreActionablesContainer = styled.div<{ isOpen?: boolean }>`
 
 export function MoreJSCollectionsMenu(props: EntityContextMenuProps) {
   const [isMenuOpen, setIsMenuOpen] = useState(false);
-  const nextEntityName = useNewJSCollectionName();
   const [confirmDelete, setConfirmDelete] = useState(false);
-
   const dispatch = useDispatch();
+
   const copyJSCollectionToPage = useCallback(
-    (actionId: string, actionName: string, pageId: string) =>
+    (actionId: string, actionName: string, pageId: string) => {
+      const nextEntityName = getJSEntityName();
       dispatch(
         copyJSCollectionRequest({
           id: actionId,
           destinationPageId: pageId,
           name: nextEntityName(`${actionName}Copy`, pageId),
         }),
-      ),
-    [dispatch, nextEntityName],
+      );
+    },
+    [dispatch],
   );
+
   const moveJSCollectionToPage = useCallback(
-    (actionId: string, actionName: string, destinationPageId: string) =>
+    (actionId: string, actionName: string, destinationPageId: string) => {
+      const nextEntityName = getJSEntityName();
       dispatch(
         moveJSCollectionRequest({
           id: actionId,
           destinationPageId,
           name: nextEntityName(actionName, destinationPageId, false),
         }),
-      ),
-    [dispatch, nextEntityName, props.pageId],
+      );
+    },
+    [dispatch],
   );
   const deleteJSCollectionFromPage = useCallback(
     (actionId: string, actionName: string) =>
